Tidy audio input code
