+++ src/ZxSpectrumAudio.h
@@ -2,6 +2,7 @@
 
 #include "hardware/gpio.h"
 #include "hardware/pwm.h"
+
 //
 // PICO_AUDIO_I2S I2S audio
 //
@@ -22,4 +23,7 @@ void zxSpectrumAudioInit();
 
 void __not_in_flash_func(zxSpectrumAudioHandler)(uint32_t vA, uint32_t vB, uint32_t vC, uint32_t buzzerSmoothed, uint32_t buzzer);
 
+#ifdef EAR_PIN
+uint32_t __not_in_flash_func(zxSpectrumReadEar)();
+#endif
